Add files via upload
Screenshots from the app prototype demonstrating how it looks. Included is the resource library, which allows teachers to upload their own resources to the platform, or search and view the resources uploaded by other teachers, as well as their own if they choose to. Also is the messaging screens including the chats and groups features. Chats allow for one-on-one messaging between teachers, and groups allow for multi-person messaging based on a particular topic. There is also the Forums feature, allowing teachers to join a particular topic and make posts about it. They can write messages and send pictures to accompany their posts. Lastly is the profile screen, allowing users to see their profile details, edit them if they choose, sign-out, and delete their profile.